+++ src/kernel/mmap.c
@@ -1,7 +1,7 @@
 #include <fudge.h>
 #include "mmap.h"
 
-struct mmap_entry *mmap_find(struct mmap_header *header, unsigned int vaddress)
+struct mmap_entry *mmap_find(struct mmap_header *header, unsigned long vaddress)
 {
 
     unsigned int i;
@@ -20,7 +20,7 @@ struct mmap_entry *mmap_find(struct mmap_header *header, unsigned int vaddress)
 
 }
 
-void mmap_initentry(struct mmap_entry *entry, unsigned int type, unsigned int paddress, unsigned int vaddress, unsigned int size, unsigned int flags, unsigned int ioaddress, unsigned int iofsize, unsigned int iomsize, unsigned int ioflags)
+void mmap_initentry(struct mmap_entry *entry, unsigned int type, unsigned long paddress, unsigned long vaddress, unsigned int size, unsigned int flags, unsigned long ioaddress, unsigned int iofsize, unsigned int iomsize, unsigned int ioflags)
 {
 
     entry->type = type;
