+++ src/kernel/mmap.h
@@ -13,11 +13,11 @@ struct mmap_entry
 {
 
     unsigned int type;
-    unsigned int paddress;
-    unsigned int vaddress;
+    unsigned long paddress;
+    unsigned long vaddress;
     unsigned int size;
     unsigned int flags;
-    unsigned int ioaddress;
+    unsigned long ioaddress;
     unsigned int iofsize;
     unsigned int iomsize;
     unsigned int ioflags;
@@ -33,6 +33,6 @@ struct mmap_header
 
 };
 
-struct mmap_entry *mmap_find(struct mmap_header *header, unsigned int vaddress);
-void mmap_initentry(struct mmap_entry *entry, unsigned int type, unsigned int paddress, unsigned int vaddress, unsigned int size, unsigned int flags, unsigned int ioaddress, unsigned int iofsize, unsigned int iomsize, unsigned int ioflags);
+struct mmap_entry *mmap_find(struct mmap_header *header, unsigned long vaddress);
+void mmap_initentry(struct mmap_entry *entry, unsigned int type, unsigned long paddress, unsigned long vaddress, unsigned int size, unsigned int flags, unsigned long ioaddress, unsigned int iofsize, unsigned int iomsize, unsigned int ioflags);
 void mmap_initheader(struct mmap_header *header);
